backend/requirements.txt
pydantic
python-dotenv

fastapi
uvicorn

groq

pypdf
